str-api/pyproject.toml
[tool.poetry]
name = "str-api"
version = "0.2.1"
description = ""
authors = ["thorwolpert <thor@wolpert.ca>"]
license = "BSD 3-Clause"
readme = "README.md"
packages = [{include = "str_api", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
flask = "2.3.3"
launchdarkly-server-sdk = "^9.0.1"
python-dotenv = "^1.0.0"
flask-sqlalchemy = "^3.0.5"
flask-migrate = "^4.0.4"
sql-versioning = { git = "https://github.com/bcgov/lear.git", subdirectory = "python/common/sql-versioning", branch = "feature-legal-name" }
sentry-sdk = "^1.24.0"
flask-babel = "^4.0.0"
flask-cors = "^4.0.0"
flask-jwt-oidc = "^0.3.0"
gunicorn = "^21.2.0"
jsonschema = {extras = ["format"], version = "^4.20.0"}
pycountry = "^23.12.11"

[tool.poetry.group.test.dependencies]
freezegun = "^1.2.2"
hypothesis = "^6.75.6"
pexpect = "^4.8.0"
pytest = "^7.2.2"
pytest-cov = "^4.0.0"
pytest-mock = "^3.10.0"
requests-mock = "^1.11.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.1.1"
pylint = "^3.0.2"
flake8 = "^6.1.0"
faker = "^23.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
